fix reload when switching tab once and for all
