+++ Sources/TMDB_MVVM_Detail/Views/Sections/MovieCreditSection.swift
@@ -1,6 +1,6 @@
 import SwiftUI
 import TMDB_Shared_Backend
-
+import TMDB_Shared_UI
 struct MovieCreditSection: View {
     let movieId: Int
     // avoiding The problem that every time the parent view redraws, a new instance of MovieCastingViewModel is created
@@ -27,27 +27,6 @@ struct MovieCreditSection: View {
         }.onFirstAppear {
             creditsViewModel.fetchMovieCredits(movieId: movieId)
         }
-    }
-}
-
-struct OnFirstAppearModifier: ViewModifier {
-    let perform: () -> Void
-    @State private var firstTime: Bool = true
-
-    func body(content: Content) -> some View {
-        content
-            .onAppear {
-                if firstTime {
-                    firstTime = false
-                    self.perform()
-                }
-            }
-    }
-}
-
-extension View {
-    func onFirstAppear(perform: @escaping () -> Void) -> some View {
-        return modifier(OnFirstAppearModifier(perform: perform))
     }
 }
